update return type of calculate to float, make it easier to upate in future when precision gets involved

+++ src/JacksonPollock/JacksonPollockBodyfatCalculator.php
@@ -20,7 +20,7 @@ class JacksonPollockBodyfatCalculator implements HealthCalculator, JacksonPolloc
         return new self($healthCalculatorOptions);
     }
 
-    public function calculate(): int
+    public function calculate(): float
     {
         if ($this->values->female()) {
             $calculator = JacksonPollockFemaleBodyDensityCalculator::class;
